- Checks for valid 5-letter word input - Implemented Backspace key behaviour
